requirements.txt
streamlit==1.31.1
pandas==2.1.1
pydeck==0.8.1b0
python-dotenv==1.0.1
numpy==1.26.1
matplotlib==3.8.0
plotly==5.22.0
